requirements.txt
langchain>=0.0.300
openai>=0.27.0
python-dotenv
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
